# server/src/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shoreline Care OS API",
    "version": "6.2.0",
    "description": "Open-source dietary, clinical nutrition, and food operations platform API for care facilities. Generated from the Express route definitions by server/scripts/generate-openapi.mjs — do not edit by hand; run `npm run docs:generate` in server/."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local Facility Gateway"
    },
    {
      "url": "https://api.shorelineops.com/api",
      "description": "Enterprise Cloud Sync"
    }
  ],
...
  "paths": {
...
    "/api/production/sheets": {
      "get": {
        "operationId": "get_production_sheets",
        "summary": "GET /api/production/sheets[?weekId=]",
        "tags": [
          "production"
        ],
        "parameters": [
          {
            "name": "weekId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "GET /api/production/sheets[?weekId=]\nproduction.read: planning roles + readonly. Servers/frontdesk/vendors excluded.\n\nAuthentication: JWT Bearer token; capabilities: production.read (admits: admin, manager, dietitian, dietary, staff, readonly).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "staff",
            "readonly"
          ],
          "capabilities": [
            "production.read"
          ]
        }
      },
      "post": {
        "operationId": "post_production_sheets",
        "summary": "POST /api/production/sheets",
        "tags": [
          "production"
        ],
        "description": "POST /api/production/sheets\nproduction.write only (no servers/frontdesk/readonly). Replay after the gate.\n\nAuthentication: JWT Bearer token; capabilities: production.write (admits: admin, manager, dietitian, dietary, staff).",
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "staff"
          ],
          "capabilities": [
            "production.write"
          ]
        }
      }
    },
    "/api/production/sheets/generate": {
      "get": {
        "operationId": "get_production_sheets_generate",
        "summary": "GET /api/production/sheets/generate?weekId=&day=&slot=",
        "tags": [
          "production"
        ],
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "slot",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          },
          {
            "name": "weekId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Query parameter read by the handler."
          }
        ],
        "description": "GET /api/production/sheets/generate?weekId=&day=&slot=\nReturns an existing sheet or auto-generates AND PERSISTS one on a miss.\nIt stays GET for frontend compatibility (src/api/production.ts getSheet),\nbut because a miss writes, it requires production.write — not just read.\n\nAuthentication: JWT Bearer token; capabilities: production.write (admits: admin, manager, dietitian, dietary, staff).",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request — validation failed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "staff"
          ],
          "capabilities": [
            "production.write"
          ]
        }
      }
    },
    "/api/production/sheets/{id}": {
      "put": {
        "operationId": "put_production_sheets_id",
        "summary": "PUT /api/production/sheets/:id",
        "tags": [
          "production"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "PUT /api/production/sheets/:id\nproduction.write only. Replay after the gate.\n\nAuthentication: JWT Bearer token; capabilities: production.write (admits: admin, manager, dietitian, dietary, staff).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SheetUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "staff"
          ],
          "capabilities": [
            "production.write"
          ]
        }
      },
      "delete": {
        "operationId": "delete_production_sheets_id",
        "summary": "DELETE /api/production/sheets/:id",
        "tags": [
          "production"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "DELETE /api/production/sheets/:id\nproduction.delete (admin only). Replay after the gate.\n\nAuthentication: JWT Bearer token; capabilities: production.delete (admits: admin).",
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin"
          ],
          "capabilities": [
            "production.delete"
          ]
        }
      }
    },
    "/api/production/sheets/{id}/signoff": {
      "post": {
        "operationId": "post_production_sheets_id_signoff",
        "summary": "POST /api/production/sheets/:id/signoff",
        "tags": [
          "production"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "POST /api/production/sheets/:id/signoff\nproduction.write only. Replay after the gate.\n\nAuthentication: JWT Bearer token; capabilities: production.write (admits: admin, manager, dietitian, dietary, staff).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/post_production_sheets_id_signoff_Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — insufficient role"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-shoreline-auth": {
          "required": true,
          "mechanism": "jwt-bearer",
          "exactRoles": [
            "admin",
            "manager",
            "dietitian",
            "dietary",
            "staff"
          ],
          "capabilities": [
            "production.write"
          ]
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/auth/login. Role claim gates requireRole ranks."
      },
      "ehrWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-EHR-Signature",
        "description": "HMAC-SHA256 of the raw body with EHR_WEBHOOK_SECRET."
      },
      "kioskApiSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Bearer <KIOSK_API_SECRET> for kiosk punch webhooks."
      },
      "stripeWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "stripe-signature",
        "description": "Stripe webhook signature with STRIPE_WEBHOOK_SECRET."
      },
      "setupSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-setup-secret",
        "description": "SETUP_BOOTSTRAP_SECRET for first-time facility initialization."
      }
    },
    "schemas": {
...
      "SheetUpdateSchema": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {}
          },
          "counts": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
...
      "post_production_sheets_id_signoff_Body": {
        "type": "object",
        "properties": {
          "staffName": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "staffName"
        ]
      },
...
    }
  },
...
}